add check before do method _v.stable

+++ src/main/java/evilcorp/todo/controller/AdminController.java
@@ -9,11 +9,13 @@ import evilcorp.todo.service.TaskService;
 import evilcorp.todo.service.UserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/admin")
@@ -30,7 +32,7 @@ public class AdminController {
     @GetMapping("/users")
     public List<UserAdminDto> showAllUsers() {
         return userService.findAllUsers().stream().map
-                (user -> new UserAdminDto(user.getId(), user.getUsername())).toList();
+                (user -> new UserAdminDto(user.getId(), user.getUsername(), user.getRole())).toList();
     }
 
     @GetMapping("/users/{userid}/tasks")
@@ -41,6 +43,11 @@ public class AdminController {
             Map<String, Object> response = new HashMap<>();
             response.put("message", "User with id " + userid + " does not exist");
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
+        }
+        if ("admin".equals(user.getRole().getName())) {
+            Map<String, Object> response = new HashMap<>();
+            response.put("message", "Cannot watch tasks of user with role 'admin'");
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(response);
         }
         return ResponseEntity.ok(new UserTasksAdminDto(user.getUsername(), tasks));
     }
@@ -56,8 +63,13 @@ public class AdminController {
         }
         if (task == null || !userid.equals(task.getUser().getId())) {
             Map<String, Object> response = new HashMap<>();
-            response.put("message", "Incorrect task id " +taskId + " for user with id " + userid);
+            response.put("message", "Incorrect task id " + taskId + " for user with id " + userid);
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(response);
+        }
+        if ("admin".equals(user.getRole().getName())) {
+            Map<String, Object> response = new HashMap<>();
+            response.put("message", "Cannot watch task of user with role 'admin'");
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(response);
         }
         return ResponseEntity.ok(new UserTaskAdminDto(user.getUsername(), task));
 
@@ -72,14 +84,47 @@ public class AdminController {
             response.put("message", "User with id " + userid + " does not exist");
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
         }
-        if(task == null  || !userid.equals(task.getUser().getId())) {
+        if (task == null || !userid.equals(task.getUser().getId())) {
             Map<String, Object> response = new HashMap<>();
-            response.put("message", "Incorrect task id " +taskId + " for user with id " + userid);
+            response.put("message", "Incorrect task id " + taskId + " for user with id " + userid);
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(response);
+        }
+        if ("admin".equals(user.getRole().getName())) {
+            Map<String, Object> response = new HashMap<>();
+            response.put("message", "Cannot delete task of user with role 'admin'");
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(response);
         }
         taskService.deleteTask(taskId);
         Map<String, Object> response = new HashMap<>();
         response.put("Deleted task", task);
+        return ResponseEntity.ok(response);
+    }
+
+    @PatchMapping("/users/{userid}/ban")
+    public ResponseEntity<Map<String, Object>> banUser(@PathVariable Long userid) {
+        String currentUsername = SecurityContextHolder.getContext().getAuthentication().getName();
+        Optional<User> currentOptionalUser = userService.findUserByUsername(currentUsername);
+        User currentUser = currentOptionalUser.get();
+        User user = userService.findUserById(userid);
+        if (user == null) {
+            Map<String, Object> response = new HashMap<>();
+            response.put("message", "User with ID " + userid + " does not exist");
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
+        }
+        if ("admin".equals(user.getRole().getName())) {
+            Map<String, Object> response = new HashMap<>();
+            response.put("message", "Cannot ban user with role 'admin'");
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(response);
+        }
+        if(user.getId().equals(currentUser.getId())) {
+            Map<String, Object> response = new HashMap<>();
+            response.put("message", "Cannot ban yourself'");
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(response);
+        }
+
+        Map<String, Object> response = new HashMap<>();
+        response.put("message", "User with ID " + userid + "and username " + user.getUsername() + " has been banned");
+        userService.deleteUser(user);
         return ResponseEntity.ok(response);
     }
